[Refactor] Divide decode Selection tracker build - Selection Tracker build 코드 분리
